--- tests/testlib/s2n_self_talk_test_utils.c
@@ -1,21 +1,0 @@
-/*
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License.
- * A copy of the License is located at
- *
- *  http://aws.amazon.com/apache2.0
- *
- * or in the "license" file accompanying this file. This file is distributed
- * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-
-#include "testlib/s2n_testlib.h"
-
-S2N_RESULT s2n_noop_inet_socket_cb(struct s2n_connection *conn)
-{
-    return S2N_RESULT_OK;
-}
