+++ tests/testlib/s2n_testlib.h
@@ -245,10 +245,3 @@ extern const s2n_parsed_extension EMPTY_PARSED_EXTENSIONS[S2N_PARSED_EXTENSIONS_
 int s2n_kem_recv_public_key_fuzz_test(const uint8_t *buf, size_t len, struct s2n_kem_params *kem_params);
 int s2n_kem_recv_ciphertext_fuzz_test(const uint8_t *buf, size_t len, struct s2n_kem_params *kem_params);
 int s2n_kem_recv_ciphertext_fuzz_test_init(const char *kat_file_path, struct s2n_kem_params *kem_params);
-
-/* A collection of callbacks run during inet socket self talk tests */
-struct self_talk_inet_socket_callbacks {
-    S2N_RESULT (*server_post_handshake_cb)(struct s2n_connection *conn);
-    S2N_RESULT (*client_post_handshake_cb)(struct s2n_connection *conn);
-};
-S2N_RESULT s2n_noop_inet_socket_cb(struct s2n_connection *conn);
